Works with Glass camera
Modified the app to take a picture with the Glass's camera and then sends that picture to the mobile device without saving it to the Glass. This allows it to not take up needless space and also acts as a privacy mechanism for those getting a picture taken of them.